requirements.txt

web3
eth-account
numpy
langchain-google-genai
langchain
python-dotenv
gunicorn
fastapi
uvicorn
pydantic
requests
